loging script failure & fixed exit on stop already stopped vm

+++ scripts/kvm/vm/stop.py
@@ -5,6 +5,10 @@ import os
 import signal
 import json
 
+if len(sys.argv) < 1:
+    sys.stderr.write('missing argument')
+    sys.exit(1)
+
 vm = json.loads(sys.argv[1])
 
 if not 'parameters' in vm or not 'pid' in vm['parameters'] or vm['parameters']['pid'] == 0:
@@ -13,6 +17,7 @@ if not 'parameters' in vm or not 'pid' in vm['parameters'] or vm['parameters']['
 pid = int(vm['parameters']['pid'])
 
 try:
-    os.kill(pid, signal.SIGTERM)
+    if not pid == 0:
+        os.kill(pid, signal.SIGTERM)
 finally:
     print "pid", 0
